refactor(core): split pack builder full delta and index update flows
